taketake: add 2nd flac encoding to the flac test
Encode the 2nd decoded wav back into another flac file and verify its
contents matches the first decoded flac.

This is a separate flac file because we can't compare against the original
flac for round-trip flac-to-wav-to-flac testing: the original wasn't
encoded with --replay-gain, and the flac algorithm may change in the
future, resulting in different binary data encoded into the test flacs vs.
the original flac checked into the repository.