taketake: fix minor check_xdelta returncode check issue
I had wrapped the ProcessLookupError try/except in an if check for the
returncode, intending to short-circuit the whole thing if the returncode
was already None.  But I got the check backwards, so we always went
through the exception handling.

But I don't think Python process handling is multithreaded, so I doubt the
exception handling is needed at all as long as we don't yield asyncio loop
control to allow the process to be reaped.