taketake: add wait() when failing check_xdelta
The small wait lets test cases finish up before trying to terminate the
xdelta3 process.  Since we only use short files in the tests, the 10ms
delay will be long enough.

For cases that the timeout is not long enough, there is a small chance that
Python will cause a warning to be emitted, depending on a race condition:

        'Unknown child process pid 266074, will report returncode 255'

See https://bugs.python.org/issue43578

Also add a bunch of tests for verifying xdelta match vs. mismatch.

And add a note that check_xdelta incorrectly can report failure for files
that are 18 bytes in size or less because xdelta3 just puts the contents of
the target file inside the xdelta in that case, meaning the source file is
not required for reconstruction.