taketake: add a short-timeout wait to check_xdelta
Was still seeing "Unknown child process" messages during tests,
so wrapping the xdelta printdelta call in an async wait with a short
timeout before going for the call to terminate.  Hopefully this lets the
really short tests finish without needing terminate, and the long tests to
take longer than the timeout (but that's unlikely).

Also track down the remaining functions still using finfo.