taketake: insert Step.autoname (untested)
To support timestamps for non-audio-timestamped files based on neighboring
files, we need a reordering step.  Might as well do this as a separate step
in the StepNetwork.  Update the design to match.

The code in Step.prompt is in an interim state, so the whole thing is
likely broken in this commit.