taketake: pyright fix for tempwavdir.is_dir
Oddly, pyright is fine with the tempwavdir.exists test right above it, but
then seems to forget that tempwavdir has to be a Path at this point.

Add comment: the monkey patching of the asyncio.subprocess.Process
stdout_data and stderr_data really is broken since it assigs them prior to
waiting for the process to finish.  This does need fixing.