taketake: 1st Stepper.cancellable implementation
This initial design provides a context manager that returns a class having
a @staticmethod that should be called at the start of execution of the body
of the statement.  If that function discovers the current token has been
canceled, it will raise the internal SkipExecution exception to prevent the
rest of the with body from executing.