taketake: allow fincore up to 1% size after flush
Add leniency to flush_fs_caches - fincore can now report up to 1% of the
file as still being cached and not fail out with a FileFlushError.

This allows for the alt-h listen/review feature to operate while flushing
for verification without killing the file transfer.  That is, mpv can
buffer up to 1% of new data out of the filesystem between the flush and
fincore completing.

This is based off of cProfile showing 6ms/GB fincore rate on an Intel i7-4770.